ID of locations should be Int64 instead of Int32
This was fixed in "Skybrud.Social" almost a year ago, so it should obviously also be fixed in this project.

+++ src/Skybrud.Social.Instagram/Endpoints/InstagramLocationsEndpoint.cs
@@ -70,7 +70,7 @@ namespace Skybrud.Social.Instagram.Endpoints {
         /// <see>
         ///     <cref>https://instagram.com/developer/endpoints/locations/#get_locations_media_recent</cref>
         /// </see>
-        public InstagramGetLocationRecentMediaResponse GetRecentMedia(int locationId) {
+        public InstagramGetLocationRecentMediaResponse GetRecentMedia(long locationId) {
             return InstagramGetLocationRecentMediaResponse.ParseResponse(Raw.GetRecentMedia(locationId));
         }
